+++ src/aiofiles/base.py
@@ -1,18 +1,64 @@
-from asyncio import get_running_loop
-from collections.abc import Awaitable
+import asyncio
+import threading
+from collections.abc import AsyncIterator, Awaitable, Callable, Coroutine
 from contextlib import AbstractAsyncContextManager
 from functools import partial, wraps
+from queue import Empty, Queue
 
 
-def wrap(func):
-    @wraps(func)
-    async def run(*args, loop=None, executor=None, **kwargs):
-        if loop is None:
-            loop = get_running_loop()
-        pfunc = partial(func, *args, **kwargs)
-        return await loop.run_in_executor(executor, pfunc)
+def to_agen(cb: Callable) -> Callable:
+    @wraps(cb)
+    async def _wrapper(*args, **kwargs) -> AsyncIterator:
+        def _iterate(
+            q: Queue, *, next_item_event: threading.Event, eoi_event: threading.Event
+        ):
+            try:
+                for row in cb(*args, **kwargs):
+                    # the event is cleared so that the iteration gets blocked
+                    # until the main generator allows the next iteration
+                    next_item_event.clear()
+                    q.put(row)
+                    next_item_event.wait()
+            finally:
+                eoi_event.set()  # end of iteration
 
-    return run
+        loop = asyncio.get_running_loop()
+        queue: Queue = Queue()  # thread-safe
+        ready_for_item = threading.Event()
+        end_of_iteration = threading.Event()
+        gen = partial(
+            _iterate,
+            q=queue,
+            next_item_event=ready_for_item,
+            eoi_event=end_of_iteration,
+        )
+        loop.run_in_executor(None, gen)
+
+        while True:
+            # in case the iterator is exhausted at this point
+            if end_of_iteration.is_set():
+                break
+            try:
+                # the `get_nowait` method is a remedy here
+                # because `queue.get()` could block the thread
+                # when queue is empty, but EOI was not set yet
+                item = queue.get_nowait()
+            except Empty:
+                continue
+            ready_for_item.set()
+            queue.task_done()
+            yield item
+        queue.join()
+
+    return _wrapper
+
+
+def wrap(cb: Callable) -> Callable:
+    @wraps(cb)
+    async def _wrapper(*args, **kwargs) -> Coroutine:
+        return await asyncio.to_thread(cb, *args, **kwargs)
+
+    return _wrapper
 
 
 class AsyncBase:
@@ -23,10 +69,9 @@ class AsyncBase:
 
     @property
     def _loop(self):
-        return self._ref_loop or get_running_loop()
+        return self._ref_loop or asyncio.get_running_loop()
 
     def __aiter__(self):
-        """We are our own iterator."""
         return self
 
     def __repr__(self):
@@ -73,7 +118,7 @@ class AiofilesContextManager(Awaitable, AbstractAsyncContextManager):
         return await self
 
     async def __aexit__(self, exc_type, exc_val, exc_tb):
-        await get_running_loop().run_in_executor(
+        await asyncio.get_running_loop().run_in_executor(
             None, self._obj._file.__exit__, exc_type, exc_val, exc_tb
         )
         self._obj = None