+++ src/aiofiles/os.py
@@ -1,40 +1,36 @@
-"""Async executor versions of file functions from the os module."""
+"""Asynchronous version of the os module."""
 
 import os
 
 from . import ospath as path
-from .base import wrap
+from .base import to_agen, wrap
 
 __all__ = [
-    "path",
-    "stat",
+    "access",
+    "getcwd",
+    "listdir",
+    "makedirs",
+    "mkdir",
+    "readlink",
+    "remove",
+    "removedirs",
     "rename",
     "renames",
     "replace",
-    "remove",
-    "unlink",
-    "mkdir",
-    "makedirs",
     "rmdir",
-    "removedirs",
-    "symlink",
-    "readlink",
-    "listdir",
+    "path",
     "scandir",
-    "access",
-    "wrap",
-    "getcwd",
+    "stat",
+    "symlink",
+    "unlink",
+    "walk",
 ]
 
 access = wrap(os.access)
-
 getcwd = wrap(os.getcwd)
-
 listdir = wrap(os.listdir)
-
 makedirs = wrap(os.makedirs)
 mkdir = wrap(os.mkdir)
-
 readlink = wrap(os.readlink)
 remove = wrap(os.remove)
 removedirs = wrap(os.removedirs)
@@ -42,12 +38,14 @@ rename = wrap(os.rename)
 renames = wrap(os.renames)
 replace = wrap(os.replace)
 rmdir = wrap(os.rmdir)
-
 scandir = wrap(os.scandir)
 stat = wrap(os.stat)
 symlink = wrap(os.symlink)
-
 unlink = wrap(os.unlink)
+# In Python 3.9:
+# 1. `inspect.isgeneratorfunction(os.walk)` is False
+# 2. `inspect.isfunction(os.walk)` is True
+walk = to_agen(os.walk)
 
 
 if hasattr(os, "link"):